split clientlist component into two modals to make code more readable, added a few adjustments to the cards, now they are responsive
