+++ src/components/BusinessDashboard/ClientsSection/EditClientModal.jsx
@@ -1,0 +1,113 @@
+import React, { useState } from 'react';
+import { FaTimes } from 'react-icons/fa';
+
+const EditClientModal = ({ client, onClose, onUpdateClient, onDeleteClient }) => {
+    const [editedClient, setEditedClient] = useState(client);
+
+    const handleInputChange = (e) => {
+        const { name, value } = e.target;
+        setEditedClient(prev => ({ ...prev, [name]: value }));
+    };
+
+    const handlePhotoUpload = (e) => {
+        const file = e.target.files[0];
+        if (file) {
+            const reader = new FileReader();
+            reader.onloadend = () => {
+                setEditedClient(prev => ({ ...prev, photo: reader.result }));
+            };
+            reader.readAsDataURL(file);
+        }
+    };
+
+    const handleUpdate = (e) => {
+        e.preventDefault();
+        onUpdateClient(editedClient);
+        onClose();
+    };
+
+    return (
+        <div className="fixed inset-0 bg-black rounded-xl bg-opacity-50 flex items-center justify-center z-50">
+            <div className="bg-white rounded-lg p-6 max-w-md w-full">
+                <div className="flex justify-between items-center mb-4">
+                    <h2 className="text-lg font-semibold">Edit Client</h2>
+                    <button onClick={onClose} className="text-gray-500 hover:text-gray-700">
+                        <FaTimes size={24} />
+                    </button>
+                </div>
+                <form onSubmit={handleUpdate}>
+                    <div className="mb-4">
+                        <label className="block text-sm font-medium text-gray-700 mb-1">
+                            Client Picture
+                        </label>
+                        <input
+                            type="file"
+                            accept="image/*"
+                            onChange={handlePhotoUpload}
+                            className="w-full text-sm text-gray-500 file:mr-4 file:py-2 file:px-4 file:rounded-full file:border-0 file:text-sm file:font-semibold file:bg-blue-50 file:text-blue-700 hover:file:bg-blue-100"
+                        />
+                    </div>
+                    <div className="mb-4">
+                        <label htmlFor="name" className="block text-sm font-medium text-gray-700 mb-1">
+                            Client Name
+                        </label>
+                        <input
+                            type="text"
+                            id="name"
+                            name="name"
+                            value={editedClient.name}
+                            onChange={handleInputChange}
+                            required
+                            className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-primary focus:border-primary"
+                        />
+                    </div>
+                    <div className="mb-4">
+                        <label htmlFor="email" className="block text-sm font-medium text-gray-700 mb-1">
+                            Client Email
+                        </label>
+                        <input
+                            type="email"
+                            id="email"
+                            name="email"
+                            value={editedClient.email}
+                            onChange={handleInputChange}
+                            required
+                            className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-primary focus:border-primary"
+                        />
+                    </div>
+                    <div className="mb-4">
+                        <label htmlFor="phone" className="block text-sm font-medium text-gray-700 mb-1">
+                            Client Phone
+                        </label>
+                        <input
+                            type="tel"
+                            id="phone"
+                            name="phone"
+                            value={editedClient.phone}
+                            onChange={handleInputChange}
+                            required
+                            className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-primary focus:border-primary"
+                        />
+                    </div>
+                    <div className="flex justify-between">
+                        <button
+                            type="submit"
+                            className="btn-blue-dashboard"
+                        >
+                            Update Client
+                        </button>
+                        <button
+                            type="button"
+                            onClick={() => onDeleteClient(client.id)}
+                            className="btn-red"
+                        >
+                            Delete Client
+                        </button>
+                    </div>
+                </form>
+            </div>
+        </div>
+    );
+};
+
+export default EditClientModal;
